Java API call script
This entire folder is for an eclipse project that makes the necessary calls that I used to build the tables. I think it's a one time use, but it could prove useful. It's very janky, so if you want to use it, definitely talk to me. The only dependency that needs to be added to the CLASSPATH is the json.jar file